Split chat_impl tests module to satisfy structural lint gate (Fixes #37)
